+++ src/com/intellij/idea/plugin/hybris/system/type/codeInsight/lookup/TSLookupElementFactory.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,6 +18,7 @@
 
 package com.intellij.idea.plugin.hybris.system.type.codeInsight.lookup
 
+import com.intellij.codeInsight.completion.PrioritizedLookupElement
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
@@ -141,4 +142,9 @@ object TSLookupElementFactory {
         .withIcon(HybrisIcons.TS_CUSTOM_PROPERTY)
         .withCaseSensitivity(false)
 
+    fun buildHeaderAbbreviation(lookupString: String) = PrioritizedLookupElement.withPriority(
+        LookupElementBuilder.create(lookupString)
+            .withTypeText("Header Abbreviation", true)
+            .withIcon(HybrisIcons.TS_HEADER_ABBREVIATION)
+        , 2.0)
 }